requirements.txt
# Core dependencies for Claude Self-Reflect
# Used by: batch automation (Docker), scripts, and runtime components

# ============================================================================
# Anthropic SDK (for Batch API)
# ============================================================================
anthropic>=0.18.0

# ============================================================================
# Vector Database
# ============================================================================
qdrant-client>=1.15.0,<2.0.0

# ============================================================================
# Embeddings
# ============================================================================
fastembed>=0.7.0,<1.0.0
voyageai>=0.3.0,<1.0.0

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ============================================================================
# Data Processing
# ============================================================================
# NumPy version varies by Python version:
# - Python <3.13: NumPy 1.x (fastembed compatible)
# - Python >=3.13: NumPy 2.x (required by fastembed 0.7.1+)
numpy>=1.24.0,<2.0.0; python_version<"3.13"
numpy>=2.1.0; python_version>="3.13"

# ============================================================================
# Utilities
# ============================================================================
# Retry and backoff
backoff>=2.2.0
tenacity>=9.1.0

# Progress bars
tqdm>=4.67.0

# Human-readable formatting
humanize>=4.13.0
rich>=13.0.0

# System utilities
psutil>=7.0.0

# HTTP requests (used in batch_monitor.py)
requests>=2.31.0

# ============================================================================
# Code Quality
# ============================================================================
# AST-GREP for static code analysis
ast-grep-py>=0.28.0

# ============================================================================
# MCP Server (optional - for development)
# ============================================================================
# Uncomment if running MCP server locally
# fastmcp>=0.0.7
# mcp-server-qdrant>=0.8.0
